Переписываю мок карточки фильма
